Merge Release/11.5.0 into master

+++ projects/components/labels/labels.component.ts
@@ -1,45 +1,56 @@
-import {Component, Input, OnChanges, HostBinding} from "@angular/core";
-import {Utils, IRef} from "@sinequa/core/base";
-import {AppService} from "@sinequa/core/app-utils";
-import {Record} from "@sinequa/core/web-services";
-import {LabelsService} from "./labels.service";
+import { Component, Input, OnChanges, HostBinding } from "@angular/core";
+import { Utils } from "@sinequa/core/base";
+import { Record } from "@sinequa/core/web-services";
+import { LabelsService } from "./labels.service";
+import { AppService } from "@sinequa/core/app-utils";
 
 @Component({
     selector: "sq-labels",
     // We need the two spans to get whitespace between each label
     // change size by adding h1-6 class to .sq-label div (default is h5)
     templateUrl: "./labels.component.html",
-    styles: [`
-.input-autocomplete{
-    display: flex;
-    flex-direction: column;
-}
-    `]
+    styles: [
+        `
+            .sq-labels-public {
+                background-color: #4fc3f7;
+            }
+            .sq-labels-private {
+                background-color: #7283a7;
+            }
+            .sq-label-remove {
+                margin-left: 3px;
+            }
+            .clickable {
+                cursor: pointer;
+            }
+            .clickable:hover {
+                opacity: 85%;
+            }
+        `,
+    ],
 })
 export class Labels implements OnChanges {
     @Input() record: Record;
     @Input() public: boolean;
-    @Input() class: string; // to preserve any classes set on host by the caller
-    @Input() field: string;
-    private labelsField: string;
+    @Input() enableDelete: boolean = false; /** Display the delete button in the label tag */
+
+    protected labelsField: string;
     showLabels: boolean;
     labels: string[];
-    newLabelRef: IRef<string>;
-    adding: boolean;
 
     @HostBinding("class") hostClasses;
 
     constructor(
         private appService: AppService,
-        private labelsService: LabelsService) {
-        this.adding = false;
-        this.newLabelRef = {value: ""};
-    }
+        private labelsService: LabelsService
+    ) {}
 
     ngOnChanges() {
-        this.labelsField = this.appService.resolveColumnAlias(this.field);
+        const field = this.public
+            ? this.labelsService.publicLabelsField
+            : this.labelsService.privateLabelsField;
+        this.labelsField = this.appService.resolveColumnAlias(field);
         this.showLabels = !!this.labelsField;
-        this.hostClasses = [this.class, this.public ? "sq-labels-public" : "sq-labels-private"].join(" ");
         this.makeLabels();
     }
 
@@ -51,8 +62,7 @@ export class Labels implements OnChanges {
         const labels = this.record[this.labelsField];
         if (Utils.isArray(labels)) {
             this.labels = this.labelsService.sort(labels.slice(), this.public);
-        }
-        else {
+        } else {
             this.labels = [];
         }
     }
@@ -64,36 +74,26 @@ export class Labels implements OnChanges {
         this.labelsService.selectLabels([label], this.public);
     }
 
-    toggleAdd() {
-        this.adding = !this.adding;
-        if (this.adding) {
-            // Reset initial value (desired?)
-            this.newLabelRef.value = "";
+    remove(index: number) {
+        if (this.canRemove()) {
+            let label = this.labels[index];
+            if (!this.public) {
+                label = <string>this.labelsService.removePrivatePrefix(label);
+            }
+            this.labelsService.removeLabels(
+                [label],
+                [this.record.id],
+                this.public
+            );
         }
     }
 
-    onSubmit(){
-        if (this.newLabelRef.value) {
-            const labels = this.labelsService.split(this.newLabelRef.value)
-                .filter((value) => this.labels.indexOf(value) === -1);
-            Utils.subscribe(this.labelsService.addLabels(labels, [this.record.id], this.public),
-                (value) => {
-                    this.adding = false;
-                });
-        }
+    canRemove(): boolean {
+        return this.public
+            ? this.enableDelete &&
+                this.labelsService.allowPublicLabelsManagement &&
+                this.labelsService.userLabelsRights &&
+                this.labelsService.userLabelsRights.canManagePublicLabels
+            : this.enableDelete && true;
     }
-
-    clickOutside = () => {
-        this.adding = false;
-    }
-
-    remove(
-        index: number) {
-        let label = this.labels[index];
-        if (!this.public) {
-            label = <string>this.labelsService.removePrivatePrefix(label);
-        }
-        this.labelsService.removeLabels([label], [this.record.id], this.public);
-    }
-
 }
